+++ bcbio/upload/__init__.py
@@ -102,8 +102,9 @@ def _get_files_variantcall(sample):
 
 def _maybe_add_validate(algorith, sample, out):
     for i, plot in enumerate(tz.get_in(("validate", "grading_plots"), sample, [])):
+        ptype = os.path.splitext(plot)[-1].replace(".", "")
         out.append({"path": plot,
-                    "type": "pdf",
+                    "type": ptype,
                     "ext": "validate%s" % ("" if i == 0 else "-%s" % (i + 1))})
     return out
 
@@ -129,9 +130,10 @@ def _maybe_add_sv(algorithm, sample, out):
                             "ext": "%s-sample" % svcall["variantcaller"],
                             "variantcaller": svcall["variantcaller"]})
             if "validate" in svcall:
-                for vkey, vext in [("csv", "csv"), ("plot", "pdf")]:
+                for vkey in ["csv", "plot"]:
                     vfile = tz.get_in(["validate", vkey], svcall)
                     if vfile:
+                        vext = os.path.splitext(vfile)[-1].replace(".", "")
                         out.append({"path": vfile,
                                     "type": vext,
                                     "ext": "%s-validate" % svcall["variantcaller"],
@@ -143,7 +145,10 @@ def _get_variant_file(x, key):
     """
     out = []
     fname = utils.get_in(x, key)
-    if fname:
+    upload_key = list(key)
+    upload_key[-1] = "do_upload"
+    do_upload = tz.get_in(tuple(upload_key), x, True)
+    if fname and do_upload:
         if fname.endswith(".vcf.gz"):
             out.append({"path": fname,
                         "type": "vcf.gz",
@@ -290,5 +295,11 @@ def _get_files_project(sample, upload_config):
         out.append({"path": sample["assembled_gtf"]})
     if dd.get_dexseq_counts(sample):
         out.append({"path": dd.get_dexseq_counts(sample)})
+    if dd.get_express_counts(sample):
+        out.append({"path": dd.get_express_counts(sample)})
+    if dd.get_express_fpkm(sample):
+        out.append({"path": dd.get_express_fpkm(sample)})
+    if dd.get_express_tpm(sample):
+        out.append({"path": dd.get_express_tpm(sample)})
 
     return _add_meta(out, config=upload_config)